requirements.txt
accelerate
bitsandbytes
diffusers
transformers>=4.45.2
huggingface_hub
hf_transfer>=0.1.8
peft>=0.13.0
decord>=0.6.0
wandb
pandas
torch
torchvision
torchao
sentencepiece>=0.2.0
imageio-ffmpeg>=0.5.1
numpy
scikit-learn
mediapy
opencv-python
imageio
einops
openpyxl
Pillow
scikit-image
Wand
scipy
pytorch_lightning==1.6.0